modified:   philo/incs/philo.h 	modified:   philo/logs/logs.c 	modified:   philo/parse/my_data_init.c 	modified:   philo/philo.c 	modified:   philo/thread/main_thread.c 	modified:   philo/thread/philo_thread.c 	modified:   philo/utils_one/allocat_philos_data.c 	modified:   philo/utils_one/creat_forks_mutex.c 	modified:   philo/utils_one/get_curr_time.c 	modified:   philo/utils_two/start_philos_cycle.c

+++ philo/utils_one/creat_forks_mutex.c
@@ -6,7 +6,7 @@
 /*   By: yrhiba <yrhiba@student.1337.ma>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/11 23:02:48 by yrhiba            #+#    #+#             */
-/*   Updated: 2023/03/12 23:32:10 by yrhiba           ###   ########.fr       */
+/*   Updated: 2023/03/14 04:57:37 by yrhiba           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ void	destroy_mutexs(t_data *data, int size)
 
 int	creat_forks_mutex(t_data *data)
 {
-	int	i;
+	LL	i;
 
 	i = 0;
 	while (i < data->number_of_philos)
