+++ philo/utils_two/start_philos_cycle.c
@@ -6,7 +6,7 @@
 /*   By: yrhiba <yrhiba@student.1337.ma>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/13 19:09:07 by yrhiba            #+#    #+#             */
-/*   Updated: 2023/03/13 19:32:56 by yrhiba           ###   ########.fr       */
+/*   Updated: 2023/03/14 04:57:09 by yrhiba           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ static void	exit_error(t_data *data, int ithr)
 
 int	start_philos_cycle(t_data *data)
 {
-	int	i;
+	LL	i;
 
 	i = 0;
 	while (i < data->number_of_philos)
